Moved canonicalize and isolated_unc to uncertainty_set - all subclasses inheirit from it.
